ai_services/requirements.txt
# AI Services 模块依赖

# 核心依赖
requests>=2.28.0
aiohttp>=3.8.0
pydantic>=1.10.0
PyYAML>=6.0

# 可选依赖 - 本地嵌入模型
sentence-transformers>=2.2.0
torch>=1.12.0
transformers>=4.20.0

# 可选依赖 - 交叉编码器重排序
scikit-learn>=1.1.0
numpy>=1.21.0

# 开发和测试依赖
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# 文档生成
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# 性能分析
memory-profiler>=0.60.0
line-profiler>=4.0.0